Dod rework

+++ Scripts/Resources/Item/ItemMining.cs
@@ -6,10 +6,10 @@ namespace TerrariaRipoffNNF;
 [GlobalClass]
 public sealed partial class ItemMining : ItemEquipment {
     [Export] public float Speed { get; private set; }
-    [Export] public float Range { get; private set; }
+    [Export] public int Range { get; private set; }
     [Export] public float Power { get; private set; }
 
-    public static ItemMining Create(float speed, float range, float power) {
+    public static ItemMining Create(float speed, int range, float power) {
         return new ItemMining {
             Speed = speed,
             Range = range,
